requirements.txt
flask # Webserver
RPi.GPIO # GPIO pin control for laser
adafruit-circuitpython-servokit # PCA9685 servo driver for pan & tilting

# Camera libraries
picamera
opencv-python

# Server deployment
gunicorn
nginx
redis